many fixes
