+++ backend/src/api/beacon/routes/index.ts
@@ -3,6 +3,7 @@ import { Router } from "express";
 import allRoute from "./all";
 import getRoute from "./get";
 import createRoute from "./create";
+import updateRoute from "./update";
 import deleteRoute from "./delete";
 import isLoggedIn from "../../middlewares/isLoggedIn";
 import isAdmin from "../../middlewares/isAdmin";
@@ -12,6 +13,7 @@ const router = Router();
 router.use("/", allRoute);
 router.use("/", getRoute);
 router.use("/", isLoggedIn, createRoute);
+router.use("/", isLoggedIn, updateRoute);
 router.use("/", isLoggedIn, isAdmin, deleteRoute);
 
 export default router;
