+++ backend/src/api/errors/errors.ts
@@ -79,5 +79,6 @@ export enum Errors {
     POST_IS_PROCESSING = "POST_IS_PROCESSING",
     INVALID_COMMENT = "INVALID_COMMENT",
     BEACON_EXISTS = "BEACON_EXISTS",
-    INVALID_BEACON = "INVALID_BEACON"
+    INVALID_BEACON = "INVALID_BEACON",
+    BEACON_NOT_FOUND = "BEACON_NOT_FOUND"
 }
