+++ frontend/src/profile/index.jsx
@@ -284,7 +284,7 @@ const Profile = () => {
                   disabled={!user || changePwBtnDisabled}
                   value={newPw}
                   onChange={e => setNewPw(e.target.value)}
-                  helperText="Minimo 8 caratteri, almeno un numero, una maiuscola e un carattere speciale"
+                  helperText="Minimo 8 caratteri, almeno un numero e una maiuscola"
                 />
               </div>
             </div>
